Move TestS3 into a submodule
